refactor: extract interface and trait
Extracts the `DEFAULT_PROCESS_NAME` constant into `RequestHandlerConstantsInterface`.
Extracts the body of `SwooleRequestHandlerRunner` into `RequestHandlerRunnerTrait`.
Updates `SwooleRequestHandlerRunner` to implement `RequestHandlerConstantsInterface` and use `RequestHandlerRunnerTrait` to define its work.
